Moves ads tests in separate file
https://github.com/brave/brave-browser/issues/7183
